+++ src/pages-flat/sign-up/sign-up-form/sign-up-form.module.scss
@@ -1,0 +1,59 @@
+@mixin margin15 {
+  margin-bottom: 1.5rem;
+}
+
+.signBlock {
+  display: flex;
+  flex-direction: column;
+
+  width: 23.5rem;
+  margin: 1.5rem auto 0;
+  padding: 1.5rem;
+
+  .title {
+    margin-bottom: 0.8125rem;
+    text-align: center;
+  }
+
+  .iconsWrapper {
+    @include margin15;
+
+    display: flex;
+    gap: 60px;
+    justify-content: center;
+  }
+
+  .username {
+    @include margin15;
+  }
+
+  .email {
+    @include margin15;
+  }
+
+  .password {
+    @include margin15;
+  }
+
+  .passwordConfirm {
+    margin-bottom: 0.75rem;
+  }
+
+  .terms {
+    @include margin15;
+
+    .link {
+      color: var(--color-accent-300);
+    }
+  }
+
+  .signUpBtn {
+    margin-bottom: 1.125rem;
+    padding: 0.38rem 0;
+  }
+
+  .subtitle {
+    margin-bottom: 0.375rem;
+    text-align: center;
+  }
+}
